Extract `anvil` into own module for reusability
